==> 07-references.Rmd <==
# References

GERDEMAN, D. 2018. The Airbnb Effect: Cheaper Rooms for Travellers, Less Revenue for Hotels. Forbes.

HOUSING AND DEVELOPMENT BOARD. 2020. Regulations for Renting Out Your Flat [Online]. Available: https://www.hdb.gov.sg/cs/infoweb/residential/renting-out-a-flat/-bedroom/renting-out-your-flat/regulations-for-renting-out-of-flats [Accessed 8 Mar 2020].

INSIDE AIRBNB. 2020. Get the Data [Online]. Available: http://insideairbnb.com/get-the-data.html [Accessed 6 Mar 2020].


NIEUWLAND, S. & VAN MELIK, R. 2020. Regulating Airbnb: how cities deal with perceived negative externalities of short-term rentals. Current Issues in Tourism, 23, 811-825.

WONG, L. 2017. The Planning (Amendment) Bill 2017 Round-Up Speech by Lawrence Wong. Singapore.
